EMBCESSMOD-5146 Fix needs form to preserve the current form updates in the step ess file wizard
